[branch-2.10] [fix] [auth] fix not forward compatible config saslJaasServerRoleTokenSignerSecretPath after cherry-pick #15121 (#19971)
After cherry-picked #15121 into branch-2.10 to solve the issue sasl authentication failure, we will do a follow-up process to keep the new configuration `saslJaasServerRoleTokenSignerSecretPath` forward compatible: make this config optinal.

+++ pulsar-broker-auth-sasl/src/test/java/org/apache/pulsar/broker/authentication/ProxySaslAuthenticationTest.java
@@ -37,6 +37,7 @@ import javax.security.auth.login.Configuration;
 import lombok.Cleanup;
 import org.apache.commons.io.FileUtils;
 import org.apache.curator.shaded.com.google.common.collect.Maps;
+import org.apache.pulsar.broker.ServiceConfiguration;
 import org.apache.pulsar.client.admin.PulsarAdmin;
 import org.apache.pulsar.client.api.Authentication;
 import org.apache.pulsar.client.api.AuthenticationFactory;
@@ -51,6 +52,7 @@ import org.apache.pulsar.client.impl.auth.AuthenticationSasl;
 import org.apache.pulsar.common.configuration.PulsarConfigurationLoader;
 import org.apache.pulsar.proxy.server.ProxyConfiguration;
 import org.apache.pulsar.proxy.server.ProxyService;
+import org.mockito.Mockito;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testng.Assert;
@@ -214,10 +216,14 @@ public class ProxySaslAuthenticationTest extends ProducerConsumerBase {
 	@Override
 	@AfterMethod(alwaysRun = true)
 	protected void cleanup() throws Exception {
-		FileUtils.deleteQuietly(brokerSecretKeyFile);
-		Assert.assertFalse(brokerSecretKeyFile.exists());
-		FileUtils.deleteQuietly(proxySecretKeyFile);
-		Assert.assertFalse(proxySecretKeyFile.exists());
+		if (brokerSecretKeyFile != null) {
+			FileUtils.deleteQuietly(brokerSecretKeyFile);
+			Assert.assertFalse(brokerSecretKeyFile.exists());
+		}
+		if (proxySecretKeyFile != null) {
+			FileUtils.deleteQuietly(proxySecretKeyFile);
+			Assert.assertFalse(proxySecretKeyFile.exists());
+		}
 		super.internalCleanup();
 	}
 
@@ -292,6 +298,16 @@ public class ProxySaslAuthenticationTest extends ProducerConsumerBase {
 		consumer.close();
 
 		proxyService.close();
+	}
+
+	@Test
+	public void testNoErrorEvenIfTheConfigSecretIsEmpty () throws Exception {
+		ServiceConfiguration configurationWithoutSecret = Mockito.spy(conf);
+		Mockito.doAnswer(invocation -> null).when(configurationWithoutSecret).getSaslJaasServerRoleTokenSignerSecretPath();
+		configurationWithoutSecret.setSaslJaasServerRoleTokenSignerSecretPath(null);
+		AuthenticationProviderSasl authenticationProviderSasl = new AuthenticationProviderSasl();
+		authenticationProviderSasl.initialize(configurationWithoutSecret);
+		authenticationProviderSasl.close();
 	}
 
 	private PulsarClient createProxyClient(String proxyServiceUrl, int numberOfConnections) throws PulsarClientException {
